AdditionalRentalItem

+++ ReCapProject/src/main/java/com/etiya/ReCapProject/business/concretes/AdditionalRentalItemManager.java
@@ -1,18 +1,19 @@
 package com.etiya.ReCapProject.business.concretes;
 
 import com.etiya.ReCapProject.business.abstracts.AdditionalRentalItemService;
+import com.etiya.ReCapProject.business.abstracts.AdditionalServiceService;
+import com.etiya.ReCapProject.business.abstracts.RentalService;
 import com.etiya.ReCapProject.business.constants.Messages;
 import com.etiya.ReCapProject.business.dtos.AdditionalRentalItemSearchListDto;
 import com.etiya.ReCapProject.business.requests.additionalRentalItemRequests.CreateAdditionalRentalItemRequest;
 import com.etiya.ReCapProject.business.requests.additionalRentalItemRequests.DeleteAdditionalRentalItemRequest;
 import com.etiya.ReCapProject.business.requests.additionalRentalItemRequests.UpdateAdditionalRentalItemRequest;
+import com.etiya.ReCapProject.core.utilities.business.BusinessRules;
 import com.etiya.ReCapProject.core.utilities.mapping.ModelMapperService;
-import com.etiya.ReCapProject.core.utilities.results.DataResult;
-import com.etiya.ReCapProject.core.utilities.results.Result;
-import com.etiya.ReCapProject.core.utilities.results.SuccessDataResult;
-import com.etiya.ReCapProject.core.utilities.results.SuccessResult;
+import com.etiya.ReCapProject.core.utilities.results.*;
 import com.etiya.ReCapProject.dataAccess.abstracts.AdditionalRentalItemDao;
 import com.etiya.ReCapProject.entities.concretes.AdditionalRentalItem;
+import com.etiya.ReCapProject.entities.concretes.AdditionalService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
@@ -25,10 +26,15 @@ public class AdditionalRentalItemManager implements AdditionalRentalItemService 
 
     private ModelMapperService modelMapperService;
     private AdditionalRentalItemDao additionalRentalItemDao;
+    private RentalService rentalService;
+    private AdditionalServiceService additionalServiceService;
+
     @Autowired
-    public AdditionalRentalItemManager(ModelMapperService modelMapperService, AdditionalRentalItemDao additionalRentalItemDao) {
+    public AdditionalRentalItemManager(ModelMapperService modelMapperService, AdditionalRentalItemDao additionalRentalItemDao, RentalService rentalService, AdditionalServiceService additionalServiceService) {
         this.modelMapperService = modelMapperService;
         this.additionalRentalItemDao = additionalRentalItemDao;
+        this.rentalService = rentalService;
+        this.additionalServiceService = additionalServiceService;
     }
 
     @Override
@@ -42,6 +48,10 @@ public class AdditionalRentalItemManager implements AdditionalRentalItemService 
     @Override
     public Result add(CreateAdditionalRentalItemRequest createAdditionalRentalItemRequest) {
 
+        var result = BusinessRules.run(isAdditionalServiceExists(createAdditionalRentalItemRequest.getAdditionalServiceId()),isRentalExists(createAdditionalRentalItemRequest.getRentalId()));
+        if (result!=null){
+            return result;
+        }
         AdditionalRentalItem additionalRentalItem=modelMapperService.forRequest().map(createAdditionalRentalItemRequest,AdditionalRentalItem.class);
         this.additionalRentalItemDao.save(additionalRentalItem);
         return new SuccessResult(Messages.ADDITIONALSERVICEADD);
@@ -49,6 +59,10 @@ public class AdditionalRentalItemManager implements AdditionalRentalItemService 
 
     @Override
     public Result delete(DeleteAdditionalRentalItemRequest deleteAdditionalRentalItemRequest) {
+        var result = BusinessRules.run(isAdditionalRentalItemExists(deleteAdditionalRentalItemRequest.getId()));
+        if (result!=null){
+            return result;
+        }
         AdditionalRentalItem additionalRentalItem = modelMapperService.forRequest()
                 .map(deleteAdditionalRentalItemRequest, AdditionalRentalItem.class);
         this.additionalRentalItemDao.delete(additionalRentalItem);
@@ -57,6 +71,10 @@ public class AdditionalRentalItemManager implements AdditionalRentalItemService 
 
     @Override
     public Result update(UpdateAdditionalRentalItemRequest updateAdditionalRentalItemRequest) {
+        var result = BusinessRules.run(isAdditionalRentalItemExists(updateAdditionalRentalItemRequest.getId()),isAdditionalServiceExists(updateAdditionalRentalItemRequest.getAdditionalServiceId()),isRentalExists(updateAdditionalRentalItemRequest.getRentalId()));
+        if (result!=null){
+            return result;
+        }
         AdditionalRentalItem additionalRentalItem = modelMapperService.forRequest()
                 .map(updateAdditionalRentalItemRequest, AdditionalRentalItem.class);
         this.additionalRentalItemDao.save(additionalRentalItem);
@@ -65,6 +83,43 @@ public class AdditionalRentalItemManager implements AdditionalRentalItemService 
 
     @Override
     public DataResult<List<AdditionalRentalItem>> getByRentalId(int rentalId) {
+        var businessResult = BusinessRules.run(isRentalExists(rentalId));
+        if (businessResult!=null){
+            return new ErrorDataResult(businessResult);
+        }
         return new SuccessDataResult<>(this.additionalRentalItemDao.getByRentalId(rentalId));
     }
+
+
+
+    private Result isRentalExists(int rentalId){
+        var result = this.rentalService.isRentalExistsById(rentalId);
+        if (!result.isSuccess()){
+            return new ErrorResult(Messages.RENTALNOTFOUND);
+        }
+        return new SuccessResult();
+    }
+
+    private Result isAdditionalServiceExists(int id){
+        var result = this.additionalServiceService.checkIsAdditionalServiceExists(id);
+        if (!result.isSuccess()){
+            return new ErrorResult(Messages.ADDITIONALSERVICENOTFOUND);
+        }
+        return new SuccessResult();
+    }
+
+    private Result isAdditionalRentalItemExists(int id){
+
+        var result = this.additionalRentalItemDao.existsById(id);
+        if (!result){
+            return new ErrorResult(Messages.ADDITIONALRENTALITEMNOTFOUND);
+        }
+        return new SuccessResult();
+    }
+
+
+
+
+
+
 }
